chore(database): duplication improvements
