+++ modules/database/src/adapters/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './validateFieldChanges';
 export * from './validateFieldConstraints';
+export * from './database-transform-utils';
